Restructure homework class.
