+++ schulcloud/HomeworkHelper.swift
@@ -38,7 +38,7 @@ public class HomeworkHelper {
             })
             .flatMap { save(privateContext: privateMOC) }
             .flatMap { _ -> FetchResult in
-                NotificationCenter.default.post(name: NSNotification.Name(rawValue: Homework.changeNotificationName), object: nil)
+                NotificationCenter.default.post(name: NSNotification.Name(rawValue: Homework.homeworkDidChangeNotificationName), object: nil)
                 return Future(value: Void())
         }
     }
